move APIError, simplify message
